# crates/api/src/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "CE-RISE Hex Core Service API",
        "version": "0.0.3",
        "description": "Public and admin API for validating, creating, and querying records against versioned CE-RISE model artifacts."
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local default server"
        }
    ],
...
    "paths": {
...
        "/admin/status": {
            "get": {
                "tags": ["Admin"],
                "operationId": "status",
                "summary": "Runtime status",
                "description": "Returns runtime status payload useful for operations dashboards.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status payload returned.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": true
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT bearer token for authenticated endpoints."
            }
        },
...
    }
}